chore(refactor): moved skipped commented usages to helper
